// HealthyAtHomeAPI/Frontend/src/enums/planOptionEnums.ts
export enum Gender {
    Male = 1,
    Female = 2,
    Other = 3,
}

export enum InventoryType {
    PullupBar = 1,
    Dumbbell = 2,
    Kettlebell = 3,
    Backpack = 4,
    Table = 5,
    Chair = 6,
    ResistanceBand = 7,
    GymnasticRings = 8,
}

export enum ExerciseType {
    HorizontalPush = 1,
    HorizontalPull = 2,
    VerticalPush = 3,
    VerticalPull = 4,
    Core = 5,
    Legs = 6,
}

export enum CueType {
    Do = 1,
    Dont = 2,
}

export enum WorkoutGoal {
    Strength = 1,
    Muscle_Growth = 2,
    Endurance = 3,
}

export enum MuscleGroup {
    Chest = 2,
    Back = 3,
    Arms = 4,
    Core = 5,
    Legs = 6,
    Shoulders = 7,
}

export enum HeightUnits {
    Centimeters = 1,
    FeetInches = 2,
}

export enum WeightUnits {
    Kilograms = 1,
    Pounds = 2,
}

export enum LevelsOfDifficulty {
    Novice = 1,
    Intermediate = 2,
    Advanced = 3,
}